transactions_stream_loading/requirements.txt
pydantic
requests
kafka-python
psycopg2-binary
python-json-logger
python-dotenv
prometheus-client

# local packages
common_config-1.0.0-py3-none-any.whl
common_logger-1.0.0-py3-none-any.whl
common_models-1.0.0-py3-none-any.whl



